cli_classic: Add option to use first detected chip if multiple found
In many cases, when flashrom detects multiple chips, it is simply unable
to distinguish between similar chips in the same family (eg, W25Q64).
In these instances, we may want to programmatically use the first chip
found, so add a parameter enabling this.

TEST=read flash on google/link with --use-first-chip option, which
previously required specifying the chip using -c.

Change-Id: I0427b1ef80e4eca16623f0fc9119d79f7dd62551
Signed-off-by: Matt DeVillier <matt.devillier@gmail.com>